requirements.txt
python-dotenv>=1.2.1
requests>=2.32.5
streamlit>=1.51.0
watchdog>=6.0.0